# src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SafetyNet Alerts App",
    "description": "SafetyNet Alerts App is design to help safety teams saving lives by providing updated data from their juridiction. This app will provide data on persons, their whereabouts when their is a danger.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9000/",
      "description": "Basepath of the api"
    }
  ],
  "paths": {
    "/firestations": {
      "post": {
        "tags": [
          "Add Firestations"
        ],
        "summary": "Add Firestations data to the database",
        "operationId": "AddFirestationToDatabase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "models_input/firestations.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "models_output/addressesRsp.schema.json"
                }
              }
            }
          }
        }
      }
    },
...
    "/firestation": {
      "get": {
        "tags": [
          "Station Number"
        ],
        "summary": "Retourne la liste des personnes couverte par la caserne de pompier correspondante.",
        "operationId": "getPersonsInfosFromFirestationID",
        "parameters": [
          {
            "name": "stationNumber",
            "in": "query",
            "description": "firestation ID",
            "required": true,
            "allowEmptyValue": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "HTTP_Responses/GET/firestation.get.schema.json"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Station Number"
        ],
        "summary": "Add a mapping firestation for a specific address.",
        "operationId": "AddFirestationMappingToSpecifiedAddress",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "address",
            "required": true,
            "allowEmptyValue": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accepted"
          }
        }
      },
      "put": {
        "tags": [
          "Station Number"
        ],
        "summary": "Update a mapping firestation for a specific address.",
        "operationId": "UpdateFirestationMappingToSpecifiedAddress",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "address",
            "required": true,
            "allowEmptyValue": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accepted"
          }
        }
      },
      "delete": {
        "tags": [
          "Station Number"
        ],
        "summary": "Delete a mapping firestation for a specific address.",
        "operationId": "DeleteFirestationMappingToSpecifiedAddress",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "address",
            "required": true,
            "allowEmptyValue": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accepted"
          }
        }
      }
    },
...
  }
}